// src/app/shared/enums/teams-tab.enum.ts
export enum TeamsUrlTab {
  CLIENT = 'clients',
  TEAMS = 'teams',
  ROLES = 'roles',
}

export enum TeamsTab {
  CLIENT_TAB = 'Manage clients & environment / service(s)',
  TEAM_TAB = 'Manage team members',
  ROLES_TAB = 'Manage roles & permissions',
}

export enum TeamMemberStatus {
  CREATED = 'created',
  ACTIVE = 'active',
  INACTIVE = 'inactive',
  DELETED = 'deleted'
}

export enum ProcessUrlTab {
  PROCESS_GROUP = 'process-group',
  PROCESS = 'process',
  ARGUMENT = 'argument',
}