// src/app/shared/enums/error-tab.enum.ts
export enum ErrorTab {
  ALL = 'All Errors',
  NEW = 'New Errors',
  UNRESOLVED_MANUAL = 'Unresolved Manual Errors',
  UNRESOLVED_JIRA = 'Unresolved Jira Errors',
  RESOLVED = 'Resolved Errors',
}

export enum ErrorUrlTab {
  ALL = 'all-errors',
  NEW = 'new-errors',
  UNRESOLVED_MANUAL = 'unresolved-manual-errors',
  UNRESOLVED_JIRA = 'unresolved-jira-errors',
  RESOLVED = 'resolved-errors',
}

export enum EmptyErrorsTableHeader {
  ERROR_TYPE = 'Error type',
  DESCRIPTION = 'Description',
  SERVICE = 'Service',
  GENERATED_ON = 'Generated on',
  ASSIGNEE = 'Assignee',
  STATUS = 'Status',
  JIRA_TICKET = 'Jira ticket',
  JIRA_ID = 'Jira ID',
}

export enum ErrorsTableHeader {
  ERROR_TYPE = 'errorType',
  DESCRIPTION = 'description',
  SERVICE = 'service',
  GENERATED_ON = 'generatedOn',
  ASSIGNEE = 'assignee',
  STATUS = 'status',
  JIRA_TICKET = 'jiraTicket',
  JIRA_ID = 'jiraId',
}

export enum ErrorFilterTitle {
  ERROR_TYPE = 'Error Type',
  STATUS = 'Status',
  ASSIGNEE = 'Assignee'
}